stats.php update
Checks for text file holding links to TF2log files. If it does not
exist, creates it. Other wise it stores the URL to the log. If a user
tries to upload a duplicate, it echos a warning. However, IT DOES NOT
STOP THE UPLOAD PROCESS. That is a change that still needs to be fixed.